% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cidx_right}
\alias{cidx_right}
\title{Function to get the vector index of the grid right to the current grid.}
\usage{
cidx_right(idx, r, c)
}
\arguments{
\item{idx}{Current locations (vector index in the grid) of cars of a certain color.}

\item{r}{numbers of rows}

\item{c}{number of columns}
}
\description{
c++ implementation of the idx_right() fucntion
}

