% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grid.R
\name{createBMLGrid}
\alias{createBMLGrid}
\title{Create a BML Grid (old version)}
\usage{
createBMLGrid(r, c, ncars)
}
\arguments{
\item{r}{Number of rows}

\item{c}{Number of columns}

\item{ncars}{Vector of length 2 where first component is the number of red cars and second component in the number of blue cars}
}
\value{
BML grid
}
\description{
Create a BML Grid (old version)
}
\examples{
createBMLGrid(10,10,c(30,40))
}

