% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BMLGrid.R
\name{plot.BMLGrid}
\alias{plot.BMLGrid}
\title{plot method for BMLGrid class object}
\usage{
\method{plot}{BMLGrid}(g, ...)
}
\arguments{
\item{g}{A BMLGrid class object.}

\item{...}{Other input arguments are simply ignored.}
}
\description{
Plot the cars on the grid as red/blue squares over a white background.
}
\examples{
library(BMLGrid)
g = createBMLGrid(r = 100, c = 99, ncars = c(red = 100, blue = 100))
plot(g)
}

