% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BMLGrid.R
\name{validateBMLGrid}
\alias{validateBMLGrid}
\title{BMLGrid class validating function}
\usage{
validateBMLGrid(g)
}
\arguments{
\item{g}{A BMLGrid class object.}
}
\description{
Validate that the 3 components of a BMLGrid class object is consistent
}
\examples{
library(BMLGrid)
g = createBMLGrid(r = 100, c = 99, ncars = c(red = 100, blue = 100))
validateBMLGrid(g)
g.out = runBMLGrid(g, numSteps = 10000)
validateBMLGrid(g.out)
}

