% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BMLGrid.R
\name{createBMLGrid}
\alias{createBMLGrid}
\title{Constructor for S3 class BMLGrid}
\usage{
createBMLGrid(r, c, ncars)
}
\arguments{
\item{r}{A non-negative integer, the number of rows of the grid.}

\item{c}{A non-negative integer, the number of columns of the grid.}

\item{ncars}{A named vector of 2 non-negative integers where \code{ncars['red']}, \code{ncars['blue']} represent the number of red/blue cars in the grid, respectively.}
}
\value{
A BMLGrid class object which is essentially a matrix.
}
\description{
Constructor for S3 class BMLGrid
}
\examples{
library(BMLGrid)
g = createBMLGrid(r = 100, c = 99, ncars = c(red = 100, blue = 100))
}

