% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{crunBMLGrid}
\alias{crunBMLGrid}
\title{Simulator for Biham-Middleton-Levine Traffic Model.}
\usage{
crunBMLGrid(g, numSteps)
}
\arguments{
\item{g}{A BMLGrid class object representing the initial state of the grid.}

\item{numSteps}{Number of moves/periods.}
}
\description{
The function that actuall runs the Biham-Middleton-Levine Traffic Model from an initial state by a given number of steps.
}
\examples{
library(BMLGrid)
g = createBMLGrid(r = 100, c = 99, ncars = c(red = 100, blue = 100))
g.out = crunBMLGrid(g, 10000)
plot(g.out)
}

