% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BMLGrid.R
\name{runBMLGrid}
\alias{runBMLGrid}
\title{Simulator for Biham-Middleton-Levine Traffic Model.}
\usage{
runBMLGrid(g, numSteps, movieName = NULL, recordSpeed = FALSE)
}
\arguments{
\item{g}{A BMLGrid class object representing the initial state of the grid.}

\item{numSteps}{Number of moves/periods.}

\item{movieName}{If specified as a non-NULL string, functions from package 'animation' will be used to record the BML process as a movie.}

\item{recordSpeed}{The flag value indicating whether to record and return the average speed of the red and blue cars ar each step.}
}
\value{
If recordSpeed is unspecified or specified as \code{FALSE}, returns a \code{BMLGrid} object representing the final state of the simulation; otherwise return a list where the first element is the final-state grid object and the 2nd and 3rd elements record the average speed of red cars and blue cars, respectively.
}
\description{
The function that actuall runs the Biham-Middleton-Levine Traffic Model from an initial state by a given number of steps.
}
\examples{
library(BMLGrid)
g = createBMLGrid(r = 100, c = 99, ncars = c(red = 100, blue = 100))
g.out = runBMLGrid(g, numSteps = 10000)
plot(g.out)
g.out = runBMLGrid(g, numSteps = 10000, movieName = 'movieBMLGrid', recordSpeed = TRUE)
plot(g.out$g)
summary(g.out$v.blue)
summary(g.out$v.red)
}

